Remove clause addition from redux
