+++ src/query-builder/components/AdvancedSearch.tsx
@@ -1,4 +1,6 @@
-import React, { FC, FormEvent, MouseEvent } from 'react';
+import React, { FC, FormEvent, MouseEvent, useState } from 'react';
+import { useHistory } from 'react-router-dom';
+
 import ClauseList from './ClauseList';
 import {
   Namespace,
@@ -7,17 +9,15 @@ import {
   SearchTermType,
   Operator,
   EvidenceDataPoint,
-} from '../../uniprotkb/types/searchTypes';
+} from '../types/searchTypes';
 import './styles/advanced-search.scss';
 import useDataApi from '../../shared/hooks/useDataApi';
 import apiUrls from '../../shared/config/apiUrls';
+import { createEmptyClause, createPreSelectedClauses } from '../utils/clause';
+import { stringify } from '../utils/queryStringProcessor';
 
 type AdvancedSearchProps = {
   queryString: string;
-  namespace: Namespace;
-  clauses: Clause[];
-  handleAdvancedSubmitClick: (event: FormEvent | MouseEvent) => void;
-  dispatchAddClause: () => void;
   handleFieldSelect: (clauseId: string, field: SearchTermType) => void;
   handleInputChange: (clauseId: string, value: string, id?: string) => void;
   handleEvidenceChange: (clauseId: string, value: string) => void;
@@ -28,13 +28,18 @@ type AdvancedSearchProps = {
   ) => void;
   handleLogicChange: (clauseId: string, value: Operator) => void;
   handleRemoveClause: (clauseId: string) => void;
-  dispatchSetPreSelectedClauses: () => void;
 };
 const AdvancedSearch: FC<AdvancedSearchProps> = (props) => {
   // Use namespace hook here when ready
   const { data: searchTermsData } = useDataApi<SearchTermType[]>(
     apiUrls.advancedSearchTerms
   );
+
+  const history = useHistory();
+
+  // To be replaced by getting it from url
+  const [clauses, setClauses] = useState<Clause[]>(createPreSelectedClauses());
+  const [namespace, setNamespace] = useState<Namespace>(Namespace.uniprotkb);
 
   // NOTE: move this to the corresponding component?
   const { data: goEvidenceData } = useDataApi<EvidenceDataPoint[]>(
@@ -49,18 +54,29 @@ const AdvancedSearch: FC<AdvancedSearchProps> = (props) => {
     return null;
   }
 
-  const { handleAdvancedSubmitClick, namespace, dispatchAddClause } = props;
-
   // Handle that better...
   const evidences = {
     [Evidence.GO]: goEvidenceData,
     [Evidence.ANNOTATION]: annotationEvidenceData,
   };
 
+  const addClause = () => {
+    setClauses((clauses) => [...clauses, createEmptyClause()]);
+  };
+
+  const handleSubmitClick = (event: FormEvent | MouseEvent) => {
+    event.preventDefault();
+    const queryString = stringify(clauses);
+    history.push({
+      pathname: '/uniprotkb',
+      search: queryString && `query=${queryString}`,
+    });
+  };
+
   return (
     <form
       className="advanced-search"
-      onSubmit={handleAdvancedSubmitClick}
+      onSubmit={handleSubmitClick}
       data-testid="advanced-search-form"
     >
       <fieldset>
@@ -74,6 +90,7 @@ const AdvancedSearch: FC<AdvancedSearchProps> = (props) => {
       <fieldset>
         <ClauseList
           {...props}
+          clauses={clauses}
           searchTerms={searchTermsData}
           evidences={evidences}
         />
@@ -84,7 +101,7 @@ const AdvancedSearch: FC<AdvancedSearchProps> = (props) => {
           id="add-field"
           className="button tertiary"
           data-testid="advanced-search-add-field"
-          onClick={dispatchAddClause}
+          onClick={addClause}
         >
           Add Field
         </button>
