+++ src/query-builder/components/EnumField.tsx
@@ -1,5 +1,5 @@
 import React, { Fragment } from 'react';
-import { SearchTermType } from '../../uniprotkb/types/searchTypes';
+import { SearchTermType } from '../types/searchTypes';
 
 type EnumFieldProps = {
   field: SearchTermType;
