+++ src/query-builder/components/TextField.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 
-import { SearchTermType } from '../../uniprotkb/types/searchTypes';
+import { SearchTermType } from '../types/searchTypes';
 
 type TextFieldProps = {
   field: SearchTermType;
