+++ src/query-builder/types/searchTypes.ts
