+++ src/query-builder/utils/clause.ts
@@ -2,12 +2,7 @@ import { v1 } from 'uuid';
 import { isEqual } from 'lodash-es';
 
 import { removeProperty } from '../../shared/utils/utils';
-import {
-  Operator,
-  Clause,
-  itemType,
-  dataType,
-} from '../../uniprotkb/types/searchTypes';
+import { Operator, Clause, itemType, dataType } from '../types/searchTypes';
 
 export const createEmptyClause = (): Clause => ({
   id: v1(),
