+++ src/shared/config/InfoMappings.tsx
@@ -4,7 +4,7 @@ import UniProtKBInfo from '../../uniprotkb/components/results/UniProtKBInfo';
 import BlastInfo from '../../tools/blast/components/BlastFormInfo';
 import AlignInfo from '../../tools/align/components/AlignFormInfo';
 
-import { Namespace } from '../../uniprotkb/types/searchTypes';
+import { Namespace } from '../../query-builder/types/searchTypes';
 import { JobTypes } from '../../tools/types/toolsJobTypes';
 
 const infoMappings: {
