+++ src/uniprotkb/state/searchInitialState.ts
@@ -1,6 +1,7 @@
 import { createPreSelectedClauses } from '../../query-builder/utils/clause';
-import { Clause, Namespace } from '../types/searchTypes';
+import { Clause, Namespace } from '../../query-builder/types/searchTypes';
 
+// This will be removed
 export type SearchState = {
   readonly clauses: Clause[];
   readonly namespace: Namespace;
